refactor(sdk): group LogErr and logger in helpers/
