+++ sdk-python/omnilogger/helpers/errors.py
